core: add penitencewithbead class
